pyproject.toml
[project]
name = "project-alpha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
    "streamlit>=1.48.0",
    "uvicorn>=0.35.0",
]
